pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Task Manager"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dj-database-url>=2.3.0",
    "django>=5.2",
    "django-bootstrap5>=25.1",
    "django-filter>=25.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "whitenoise>=6.9.0",
]

[dependency-groups]
dev = [
    "ruff>=0.11.4",
]
